requirement.txt
torch==1.10.0+cu111
torchvision==0.11.1+cu111
Pillow==8.4.0
tensorboard==2.8.0
matplotlib==3.5.1
numpy==1.21.2
scipy==1.7.1
opencv-python==4.5.4
